Tidy up models and move to folder to declutter a little.
